Correcting job submission file error when given externally
It was assumed that the job submission file is always part of the
package filesystem, while, sometimes users can provide their own job
submission scritps. This is now corrected, and users can use whatever
job submission script they want.

For now, only SLURM is provided to the users. Apparently, in the
near-future, multiple schedulers to be tested.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>
Signed-off-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>